requirements.txt
asgiref==3.9.1
Django==5.0.7
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
packaging==25.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2